src/api/gis/requirements.txt
attrs==23.1.0
blinker==1.7.0
click==8.1.7
colorama==0.4.6
Flask==3.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.3
psycopg2-binary==2.9.9
python-dotenv==1.0.0
referencing==0.32.0
rpds-py==0.15.2
Werkzeug==3.0.1
